Add files via upload
This is an Object-Oriented Programming course group project implemented using Swing.

Group Members:
1)K M Monoarul Islam Shovon (Batch 221)
2)Faysal Hossain Tomal (Batch 221)

This project is a Travel Management System where users can:
-> Explore tourist spots
-> Explore hotels
-> Access many other features related to travel management

The project idea was taken from the YouTube channel: 'Code for Interview'.

The project is implemented in Java. All the necessary files are included in the provided rar file. You can download the files and run the project in NetBeans or any other IDE.